More regression checks

+++ src/fmpz_mod_mpoly/get_term_exp_ui.c
@@ -17,7 +17,7 @@ void fmpz_mod_mpoly_get_term_exp_ui(ulong * exp, const fmpz_mod_mpoly_t A,
 {
     slong N = mpoly_words_per_exp(A->bits, ctx->minfo);
 
-    if (i >= (ulong) A->length)
+    if (i >= A->length)
         flint_throw(FLINT_ERROR, "fmpz_mod_mpoly_get_term_exp_ui: index out of range");
 
     mpoly_get_monomial_ui(exp, A->exps + N*i, A->bits, ctx->minfo);
