+++ src/fmpz_mod_mpoly/set_term_coeff_fmpz.c
@@ -15,7 +15,7 @@
 void fmpz_mod_mpoly_set_term_coeff_fmpz(fmpz_mod_mpoly_t A,
                       slong i, const fmpz_t c, const fmpz_mod_mpoly_ctx_t ctx)
 {
-    if (i >= (ulong) A->length)
+    if (i >= A->length)
         flint_throw(FLINT_ERROR, "fmpz_mod_mpoly_set_term_coeff_fmpz: index is out of range");
 
     fmpz_mod_set_fmpz(A->coeffs + i, c, ctx->ffinfo);
@@ -24,7 +24,7 @@ void fmpz_mod_mpoly_set_term_coeff_fmpz(fmpz_mod_mpoly_t A,
 void fmpz_mod_mpoly_set_term_coeff_ui(fmpz_mod_mpoly_t A,
                               slong i, ulong c, const fmpz_mod_mpoly_ctx_t ctx)
 {
-    if (i >= (ulong) A->length)
+    if (i >= A->length)
         flint_throw(FLINT_ERROR, "fmpz_mod_mpoly_set_term_coeff_ui: index is out of range");
 
     fmpz_mod_set_ui(A->coeffs + i, c, ctx->ffinfo);
@@ -33,7 +33,7 @@ void fmpz_mod_mpoly_set_term_coeff_ui(fmpz_mod_mpoly_t A,
 void fmpz_mod_mpoly_set_term_coeff_si(fmpz_mod_mpoly_t A,
                               slong i, slong c, const fmpz_mod_mpoly_ctx_t ctx)
 {
-    if (i >= (ulong) A->length)
+    if (i >= A->length)
         flint_throw(FLINT_ERROR, "fmpz_mod_mpoly_set_term_coeff_si: index is out of range");
 
     fmpz_mod_set_si(A->coeffs + i, c, ctx->ffinfo);
