+++ src/fq_nmod_mpoly/content_vars.c
@@ -138,7 +138,7 @@ int fq_nmod_mpoly_content_vars(
 
     for (i = 0; i < num_vars; i++)
     {
-        if (vars[i] >= (ulong) ctx->minfo->nvars)
+        if (vars[i] >= ctx->minfo->nvars)
             flint_throw(FLINT_ERROR, "fq_nmod_mpoly_content_vars: variable out of range");
     }
